requirements.txt
cx_Freeze==8.3.0
numpy==2.2.6
pdf2image==1.17.0
opencv_python==4.12.0.88
pillow==11.3.0
pywin32==311
tqdm==4.67.1